+++ src/pages/PostsManagerPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
-import { Button, Card, Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea } from "../shared/ui";
-import { Post, Comment, Tag, fetchTags, createPost, updatePost, deletePost, UserDetail } from "../entities";
+import { Card } from "../shared/ui";
+import { Post, Comment, Tag, fetchTags, deletePost, UserDetail } from "../entities";
 
 import { fetchCommentsByPostId } from "../features";
 import { PostsManager } from "../widgets/PostsManager/ui/PostsManager";
@@ -10,6 +10,8 @@ import { ModalProvider } from "../shared";
 import { PostDetailDialog } from "../widgets/ui/PostDetailDialog";
 import { CommentModifyDialog } from "../widgets/ui/CommentModifyDialog";
 import { CommentAddDialog } from "../widgets/ui/CommentAddDialog";
+import { PostModifyDialog } from "../widgets/ui/PostModifyDialog";
+import { PostAddDialog } from "../widgets/ui/PostAddDialog";
 
 const PostsManagerPage = () => {
   const navigate = useNavigate();
@@ -25,8 +27,6 @@ const PostsManagerPage = () => {
   const [selectedPost, setSelectedPost] = useState<Post | null>(null);
   const [sortBy, setSortBy] = useState(queryParams.get("sortBy") || "");
   const [sortOrder, setSortOrder] = useState(queryParams.get("sortOrder") || "asc");
-  const [showAddDialog, setShowAddDialog] = useState(false);
-  const [showEditDialog, setShowEditDialog] = useState(false);
   const [newPost, setNewPost] = useState({ title: "", body: "", userId: 1 });
   const [tags, setTags] = useState<Tag[]>([]);
   const [selectedTag, setSelectedTag] = useState(queryParams.get("tag") || "");
@@ -62,30 +62,6 @@ const PostsManagerPage = () => {
     }
   };
 
-  // 게시물 추가
-  // shared/lib/fetch
-  const addPost = async () => {
-    try {
-      const data = await createPost({ newPost });
-      setPosts([data, ...posts]);
-      setShowAddDialog(false);
-      setNewPost({ title: "", body: "", userId: 1 });
-    } catch (error) {
-      console.error("게시물 추가 오류:", error);
-    }
-  };
-
-  // 게시물 업데이트
-  const setUpdatedPost = async () => {
-    try {
-      const data = await updatePost({ post: selectedPost! });
-      setPosts(posts.map((post) => (post.id === data.id ? data : post)));
-      setShowEditDialog(false);
-    } catch (error) {
-      console.error("게시물 업데이트 오류:", error);
-    }
-  };
-
   // 게시물 삭제
   const removePost = async (id: Post["id"]) => {
     try {
@@ -115,7 +91,6 @@ const PostsManagerPage = () => {
 
   const openPostEdit = (post: Post) => {
     setSelectedPost(post);
-    setShowEditDialog(true);
   };
 
   useEffect(() => {
@@ -146,7 +121,6 @@ const PostsManagerPage = () => {
       <Card className="w-full max-w-6xl mx-auto">
         <PostsManager
           posts={posts}
-          setShowAddDialog={setShowAddDialog}
           searchQuery={searchQuery}
           setPosts={setPosts}
           setTotal={setTotal}
@@ -169,57 +143,10 @@ const PostsManagerPage = () => {
           removePost={removePost}
           openPostEdit={openPostEdit}
         />
-        {/* 게시물 추가 대화상자 */}
-        <Dialog open={showAddDialog} onOpenChange={setShowAddDialog}>
-          <DialogContent>
-            <DialogHeader>
-              <DialogTitle>새 게시물 추가</DialogTitle>
-            </DialogHeader>
-            <div className="space-y-4">
-              <Input
-                placeholder="제목"
-                value={newPost.title}
-                onChange={(e) => setNewPost({ ...newPost, title: e.target.value })}
-              />
-              <Textarea
-                rows={30}
-                placeholder="내용"
-                value={newPost.body}
-                onChange={(e) => setNewPost({ ...newPost, body: e.target.value })}
-              />
-              <Input
-                type="number"
-                placeholder="사용자 ID"
-                value={newPost.userId}
-                onChange={(e) => setNewPost({ ...newPost, userId: Number(e.target.value) })}
-              />
-              <Button onClick={addPost}>게시물 추가</Button>
-            </div>
-          </DialogContent>
-        </Dialog>
 
-        {/* 게시물 수정 대화상자 */}
-        <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
-          <DialogContent>
-            <DialogHeader>
-              <DialogTitle>게시물 수정</DialogTitle>
-            </DialogHeader>
-            <div className="space-y-4">
-              <Input
-                placeholder="제목"
-                value={selectedPost?.title || ""}
-                onChange={(e) => setSelectedPost({ ...selectedPost!, title: e.target.value })}
-              />
-              <Textarea
-                rows={15}
-                placeholder="내용"
-                value={selectedPost?.body || ""}
-                onChange={(e) => setSelectedPost({ ...selectedPost!, body: e.target.value })}
-              />
-              <Button onClick={setUpdatedPost}>게시물 업데이트</Button>
-            </div>
-          </DialogContent>
-        </Dialog>
+        <PostAddDialog posts={posts} setPosts={setPosts} newPost={newPost} setNewPost={setNewPost} />
+
+        <PostModifyDialog post={selectedPost!} setSelectedPost={setSelectedPost} posts={posts} setPosts={setPosts} />
 
         <CommentAddDialog newComment={newComment} setNewComment={setNewComment} setComments={setComments} />
